[systemtest] Tests for NetworkPolicy enhancements (#4085)
* add tests for np enhancements, create ST for NPs etc.


* Jakub's comment


* Jakub's comment vol.2
